+++ pallets/cash/src/tests/mod.rs
@@ -3,10 +3,10 @@
 pub use codec::{Decode, Encode};
 pub use frame_support::{assert_err, assert_ok, dispatch::DispatchError};
 pub use hex_literal::hex;
+pub use our_std::convert::TryInto;
+pub use our_std::{iter::FromIterator, str::FromStr};
 pub use sp_core::crypto::AccountId32;
 pub use sp_core::offchain::testing;
-
-pub use our_std::{iter::FromIterator, str::FromStr};
 
 pub mod assets;
 pub mod common;
